Security: Unescaped user input on error messages
Some validation error messages such as FieldNotFoundError() and custom
validators may contain user input as they replayed the user input back
to the browser. This patch fixes that so that validation messages are
escaped by default using markupsafe. If you want HTML in error messages
to be treated as HTML, you will need to wrap them explicitly with markupsafe.